Fix clippy errors

+++ src/business_logic/execution/objects.rs
@@ -146,8 +146,8 @@ impl CallInfo {
             for ordered_msg in call.l2_to_l1_messages {
                 let l2tol1msg =
                     L2toL1MessageInfo::new(ordered_msg.clone(), call.caller_address.clone());
-                starknet_events.remove(ordered_msg.order as usize - 1);
-                starknet_events.insert(ordered_msg.order as usize - 1, Some(l2tol1msg));
+                starknet_events.remove(ordered_msg.order - 1);
+                starknet_events.insert(ordered_msg.order - 1, Some(l2tol1msg));
             }
         }
